+++ BovineLabs.Event.Tests/Containers/NativeEventStreamTests.cs
@@ -2,6 +2,8 @@
 //     Copyright (c) BovineLabs. All rights reserved.
 // </copyright>
 
+#if BOVINELABS_TESTING_ENABLED
+
 namespace BovineLabs.Event.Tests.Containers
 {
     using System;
@@ -570,3 +572,5 @@ namespace BovineLabs.Event.Tests.Containers
         }
     }
 }
+
+#endif