+++ BovineLabs.Event.Tests/Utility/EventUtilityTests.cs
@@ -62,6 +62,47 @@ namespace BovineLabs.Event.Tests.Utility
                 es.GetEventReaders<TestEvent>(default, out _).Complete();
             }
         }
+
+        /// <summary> Tests that capacity is set to number of events. </summary>
+        [Test]
+        public void GetEventCount()
+        {
+            const int firstEventCount = 5;
+            const int secondEventCount = 3;
+
+            using (var count = new NativeArray<int>(1, Allocator.TempJob))
+            {
+                var es = this.World.GetOrCreateSystem<TestEventSystem>();
+
+                var events1 = es.CreateEventWriter<TestEvent>();
+
+                // Write some event data
+                for (var i = 0; i < firstEventCount; i++)
+                {
+                    events1.Write(i);
+                }
+
+                es.AddJobHandleForProducer<TestEvent>(default);
+
+                var events2 = es.CreateEventWriter<TestEvent>();
+
+                // Write some event data
+                for (var i = 0; i < secondEventCount; i++)
+                {
+                    events2.Write(i);
+                }
+
+                es.AddJobHandleForProducer<TestEvent>(default);
+
+                var handle = es.Ex<TestEvent>().GetEventCount(default, count);
+                handle.Complete();
+
+                Assert.AreEqual(firstEventCount + secondEventCount, count[0]);
+
+                // Make sure it doesn't block getting readers
+                es.GetEventReaders<TestEvent>(default, out _).Complete();
+            }
+        }
     }
 }